Unify Visitor implementation class
